requirements.txt
# Core dependencies
python-dotenv==1.0.1
openai==1.12.0

# LangChain and related
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-text-splitters==0.3.6
pydantic>=2.7.4,<3.0.0

# Oracle Cloud Infrastructure
oci==2.146.0

# Testing
pytest==8.0.1

# Optional utilities
typing-extensions==4.9.0

# Add these dependencies
snowflake-connector-python==3.7.0
oracledb==2.0.0
google-cloud-bigquery>=3.0.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
sqlglot==26.25.3

# FastAPI stack
fastapi==0.115.12
uvicorn[standard]==0.29.0
python-multipart==0.0.20
